Migrated FieldError to typescript
