+++ src/components/Check/Check.jsx
@@ -21,7 +21,7 @@ import { InvalidAlert } from '../FieldLine/InvalidAlert';
 import { ValidatingSpinner } from '../FieldLine/ValidatingSpinner';
 import { InfoAddonButton } from '../FieldLine/InfoAddonButton';
 import { InfoAlert } from '../FieldLine/InfoAlert';
-import FieldError from '../FieldLine/FieldError';
+import { FieldError } from '../FieldLine/FieldError';
 
 /**
  * Component for displaying bootstrap
