--- src/components/FieldLine/FieldError/index.js
@@ -1,1 +1,0 @@
-export { default } from './FieldError';
