+++ src/components/FieldLine/FieldError/index.ts
@@ -1,0 +1,2 @@
+export * from './FieldError';
+export * from './FieldError.types';
