+++ src/components/FieldLine/FieldLine.jsx
@@ -17,7 +17,7 @@ import { ValidatingSpinner } from './ValidatingSpinner';
 import { FieldLineAddon } from './FieldLineAddon';
 import { InfoAddonButton } from './InfoAddonButton';
 import { InfoAlert } from './InfoAlert';
-import FieldError from './FieldError';
+import { FieldError } from './FieldError';
 
 /**
  * Component for displaying bootstrap
